app/requirements.txt
fastapi
uvicorn
pydantic
streamlit
psycopg2-binary
python-dotenv
requests
crewai
langgraph
